ai-server/requirements.txt
fastapi==0.111.0
uvicorn==0.30.6
httpx==0.27.0
redis==5.0.7
pydantic==2.8.2
python-dotenv==1.0.1
openai==1.35.7
SQLAlchemy==2.0.32
requests==2.32.3
